pyproject.toml
[project]
name = "illyrion"
version = "0.1.0"
description = " Discord Bot for Personal use"
requires-python = ">=3.13"
dependencies = [
    "py-cord",
 "python-dotenv",
 "audioop-lts",
 "psutil>=7.0.0",
]

authors = [{ name = "Xodium", email = "info@xodium.org" }]
license = "AGPL-3.0"
license-file = "LICENSE.md"
readme = "README.md"
keywords = ["discord", "bot", "personal"]
homepage = "https://xodium.org"
repository = "https://github.com/XodiumSoftware/Illyrion"
